# docs/swagger.yaml
basePath: /
definitions:
  handler.authUserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      display_name:
        example: User Name
        type: string
      email:
        example: user@gmail.com
        type: string
      email_verified_at:
        format: date-time
        type: string
      id:
        example: 1
        type: integer
      last_login_at:
        format: date-time
        type: string
    type: object
  handler.googleLoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: refresh_token_value
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/handler.authUserResponse'
    type: object
  handler.refreshTokenRequest:
    properties:
      refresh_token:
        example: refresh_token_value
        type: string
    type: object
  handler.refreshTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: refresh_token_value
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/handler.authUserResponse'
    type: object
  handler.createParamExerciseRequest:
    properties:
      icon_url:
        example: https://example.com/icon.png
        type: string
      name:
        example: Push-up
        type: string
    type: object
  handler.createParamExerciseResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  handler.paramExerciseResponse:
    properties:
      icon_url:
        example: https://example.com/icon.png
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Push-up
        type: string
    type: object
  presenter.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Simple REST API for Workout Ledger.
  title: Workout Ledger API
  version: "1.0"
paths:
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new access token
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.refreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.refreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/presenter.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      summary: Refresh access token
      tags:
      - auth
  /v1/auth/google/callback:
    get:
      description: Exchanges the OAuth code for user info and logs in the user
      parameters:
      - description: OAuth code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.googleLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/presenter.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      summary: Handle Google login callback
      tags:
      - auth
  /v1/auth/google/login:
    get:
      description: Redirects the user to Google OAuth login
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      summary: Start Google login
      tags:
      - auth
  /v1/param_exercises:
    get:
      description: Retrieves all param exercises
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.paramExerciseResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      security:
      - BearerAuth: []
      summary: Get param exercises
      tags:
      - param_exercises
    post:
      consumes:
      - application/json
      description: Creates a new param exercise
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createParamExerciseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createParamExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/presenter.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      security:
      - BearerAuth: []
      summary: Create param exercise
      tags:
      - param_exercises
  /v1/param_exercises/{id}:
    get:
      description: Retrieves a param exercise by ID
      parameters:
      - description: Param exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.paramExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/presenter.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.APIError'
      security:
      - BearerAuth: []
      summary: Get param exercise by ID
      tags:
      - param_exercises
securityDefinitions:
  BearerAuth:
    description: Authorization header; you may paste the token directly or use "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"